+++ api/send-email.js
@@ -16,6 +16,10 @@ import {
   subscriptionWelcomeEmail,
   subscriptionRenewedEmail,
   subscriptionCanceledEmail,
+  affiliateApplicationEmail,
+  affiliateApprovedEmail,
+  affiliateAdminNotifyEmail,
+  affiliatePayoutEmail,
 } from '../src/lib/email-templates.js';
 
 const resend = new Resend(process.env.RESEND_API_KEY);
@@ -66,6 +70,18 @@ export async function sendEmail(type, to, data = {}) {
       break;
     case 'SUBSCRIPTION_CANCELED':
       tmpl = subscriptionCanceledEmail(data);
+      break;
+    case 'AFFILIATE_APPLICATION':
+      tmpl = affiliateApplicationEmail(data);
+      break;
+    case 'AFFILIATE_APPROVED':
+      tmpl = affiliateApprovedEmail(data);
+      break;
+    case 'AFFILIATE_ADMIN_NOTIFY':
+      tmpl = affiliateAdminNotifyEmail(data);
+      break;
+    case 'AFFILIATE_PAYOUT':
+      tmpl = affiliatePayoutEmail(data);
       break;
     default:
       throw new Error(`Unknown email type: ${type}`);